requirements.txt
matplotlib==3.9.2
moviepy==1.0.3
numpy==2.1.1
Pillow==10.4.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
